For some reason, having bytes_per_vox/bytes_total as size_t (for chunking) was breaking I/O Have reverted, but this should be fixed properly

+++ include/mri.h
@@ -8,8 +8,8 @@
  * Original Author: Bruce Fischl
  * CVS Revision Info:
  *    $Author: rge21 $
- *    $Date: 2010/01/27 19:21:45 $
- *    $Revision: 1.391 $
+ *    $Date: 2010/01/27 20:25:03 $
+ *    $Revision: 1.392 $
  *
  * Copyright (C) 2002-2007,
  * The General Hospital Corporation (Boston, MA). 
@@ -149,11 +149,11 @@ typedef struct
   // volume is allocated one big buffer.
   int    ischunked; // 1 means alloc is one big chunk
   void   *chunk; // pointer to the one big chunk of buffer
-  size_t    bytes_per_vox; // # bytes per voxels
-  size_t    bytes_per_row; // # bytes per row
-  size_t    bytes_per_slice; // # bytes per slice
-  size_t    bytes_per_vol; // # bytes per volume/timepoint
-  size_t    bytes_total; // # total number of pixel bytes in the struct
+  int    bytes_per_vox; // # bytes per voxels
+  int    bytes_per_row; // # bytes per row
+  int    bytes_per_slice; // # bytes per slice
+  int    bytes_per_vol; // # bytes per volume/timepoint
+  int    bytes_total; // # total number of pixel bytes in the struct
   COLOR_TABLE *ct ;
 }
 MRI_IMAGE, MRI ;
